Decouple daemon node from daemon
They should not share anything.
